Move an integration test out of PreparedStatementTests
Move an integration test that requires a different ccm cluster
out of PreparedStatementTests into their own fixture. I've assigned
the category long to the new test as this functionality
is already covered in PrepareSimulatorTests.
